+++ CHANGELOG.md
@@ -5,6 +5,15 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 [Keep a CHANGELOG](http://keepachangelog.com).
 
 
+## Unreleased
+
+- Use the new `RDF.Literal.matches?/3` function from RDF.ex 0.6.1 for the `REGEX` function
+
+
+[Compare v0.3.0...HEAD](https://github.com/marcelotto/sparql-ex/compare/v0.3.0...HEAD)
+
+
+
 ## 0.3.0 - 2019-04-06
 
 ### Changed